17 days
17days

<신규 작업>
4. 사용자관리 메뉴 만들기 -진행중
4-3 사용자관리 수정 -완료 (역할항목, 구분항목이 수정안됨 내일확인)
4-4 사용자관리 삭제 -완료
4-5 사용자관리 passwd및 사용여부 처리 -예정
4-6 데이터리셋메뉴 사용자관리메뉴 셋팅하기 -예정
5. 로그인 기능 만들기 -예정

<미룬 일>

1. 전역으로 이쁜 폰트하기 (개선) ★
4. 메인그리드의 t헤더를 고정하고 스크롤 생기게하기 (개선) ★★★
5. 리사이저바가 부드럽게 작용이 안됨 (버그) ★★★
6. 메인컨텐츠의 div 빨간영역선 지우고 div의 백그라운드색혹은 보더 디자인 꾸미기 (개선)★
7. 검색필터에 검색을 하면 페이지당 줄수와 페이지번호가 안맞음(버그) ★★★★★ -완료
8. 역할관리메뉴 신수삭 후 메인그리드 데이터갱신 안됨(버그) ★★★★★ -완료
9. 신규시 해당 테이블의 max가 아닌 시퀀스테이블을생성해서 넣기(버그) -완료
10. 신규,수정삭제시 생성자, 변경자 클라이언트의 세션 user_id를 받아서 넣기(버그) - 아직못함
11. 화면의 div값이 동적으로 변경될때도 사이즈는 고정값을 유지하게 처리(버그)
12.수정,삭제버튼클릭시 선택이 없어도 모달창이 나옴. 알럿창이 나오게처리(버그) -완료
14. DDL에 유니크값, 디폴트값 제약조건 넣기
15. DB를 회사 VPN에 놓기 -완료

+++ src/main/java/com/kdg/api/controller/admin/MenuMenegementController.java
@@ -36,7 +36,7 @@ public class MenuMenegementController {
             // 상위 메뉴를 MenuTreeDTO로 변환
             for (MenuDTO upMenu : upMenuList_) {
                 // 하위 메뉴를 MenuTreeDTO로 변환하여 상위 메뉴에 추가
-                List<Integer> intList = new ArrayList<>();
+                List<Long> intList = new ArrayList<>();
                 List<String> stringList = new ArrayList<>();
                 for (MenuDTO downMenu : downMenuList_) {
                     if (downMenu.getOrg_menu_id() == upMenu.getMenu_id()) {
@@ -87,7 +87,7 @@ public class MenuMenegementController {
     public Object updateMenu(@PathVariable("menu_id") Long menuId, @RequestBody MenuDTO menuDTO) {
         try {
             // 게시판 정보 수정
-            menuDTO.setMenu_id(menuId.intValue());
+            menuDTO.setMenu_id((long) menuId.intValue());
             menuMenegementService.updateMenuData(menuId, menuDTO);
             // 응답 데이터 구성
             Object response = new Object() {
