+++ src/main/java/com/kdg/api/model/MenuTreeDTO.java
@@ -5,23 +5,23 @@ import java.util.List;
 
 public class MenuTreeDTO {
 
-    private int top_menu_id;
+    private Long top_menu_id;
     private String top_menu_name;
-    private List<Integer> child_menu_id_list;
+    private List<Long> child_menu_id_list;
     private List<String> child_munu_name_list;
 
     // 생성자
     public MenuTreeDTO() {
     }
 
-    public MenuTreeDTO(int menu_id, String menu_name,List<Integer> menu_id_list, List<String> menu_name_list) {
+    public MenuTreeDTO(Long menu_id, String menu_name,List<Long> menu_id_list, List<String> menu_name_list) {
         this.top_menu_id=menu_id;
         this.top_menu_name=menu_name;
         this.child_menu_id_list=menu_id_list;
         this.child_munu_name_list=menu_name_list;
     }
 
-    public void setTop_menu_id(int top_menu_id) {
+    public void setTop_menu_id(Long top_menu_id) {
         this.top_menu_id = top_menu_id;
     }
 
@@ -29,7 +29,7 @@ public class MenuTreeDTO {
         this.top_menu_name = top_menu_name;
     }
 
-    public void setChild_menu_id_list(List<Integer> child_menu_id_list) {
+    public void setChild_menu_id_list(List<Long> child_menu_id_list) {
         this.child_menu_id_list = child_menu_id_list;
     }
 
@@ -37,7 +37,7 @@ public class MenuTreeDTO {
         this.child_munu_name_list = child_munu_name_list;
     }
 
-    public int getTop_menu_id() {
+    public Long getTop_menu_id() {
         return top_menu_id;
     }
 
@@ -45,7 +45,7 @@ public class MenuTreeDTO {
         return top_menu_name;
     }
 
-    public List<Integer> getChild_menu_id_list() {
+    public List<Long> getChild_menu_id_list() {
         return child_menu_id_list;
     }
